Add TLS configuration for PostgreSQL
